[refactor] response 수정, response model 추가(wine)
