+++ app/src/main/java/com/mashup/zuzu/data/response/GetDrinksWithCategoryResponse.kt
@@ -1,36 +1,16 @@
 package com.mashup.zuzu.data.response
 
 import com.google.gson.annotations.SerializedName
+import com.mashup.zuzu.data.response.model.Wine
 
 /**
  * @Created by 김현국 2022/08/08
  */
 data class GetDrinksWithCategoryResponse(
 
-    @SerializedName("id")
-    val id: Long,
+    @SerializedName("totalPageCount")
+    val totalPageCount: Int,
 
-    @SerializedName("createdAt")
-    val createdAt: String,
-
-    @SerializedName("updatedAt")
-    val updatedAt: String,
-
-    @SerializedName("deletedAt")
-    val deletedAt: String?,
-
-    @SerializedName("name")
-    val name: String,
-
-    @SerializedName("abv")
-    val alc: Float,
-
-    @SerializedName("origin")
-    val origin: String,
-
-    @SerializedName("description")
-    val description: String,
-
-    @SerializedName("image_url")
-    val image_url: String
+    @SerializedName("list")
+    val wineList: List<Wine>
 )
